add support for marshaling and unmarshaling the LIFX LAN message header
This change adds the ability to marshal and unmarshal the Header component of a LIFX message package. The Header includes things like the target, how the target should process the message and respond, as well as the type of payload. This does not include code to marshal the packet payload.