add fmt.Stringer to PacketComponent interface and implement methods
To make it easier to understand the contents of each struct, I've added `fmt.Stringer` to the `lifxprotocol.PacketCompnent` interface. This allows us to print out each struct, and have a meaningful string representation of its contents. This is meant primarily for development and debugging purposes as well as logging (maybe?).